+++ UnitTests/UnitTest_TextureAnimations/FadeNodeAnimator.cpp
@@ -88,7 +88,7 @@ namespace demo
 					float speed = 1.0f;
 					if (mFadeType == 1) // random not-squares
 					{
-						speed = mSpeed[index % (getCountX() * getCountY())];
+						speed = (float)mSpeed[index % (getCountX() * getCountY())];
 					}
 					if (mFadeType == 2) // from center
 					{
